change some identifiers from u_int64_t to unsigned long long
makes printf much easier on 64bit platforms...

+++ openbsc/include/openbsc/gsm_data.h
@@ -416,7 +416,7 @@ struct gsm_network {
 #define SMS_HDR_SIZE	128
 #define SMS_TEXT_SIZE	256
 struct gsm_sms {
-	u_int64_t id;
+	unsigned long long id;
 	struct gsm_subscriber *sender;
 	struct gsm_subscriber *receiver;
